requirements.txt
flask
pandas
scikit-learn
numpy
python-dotenv
joblib
google-genaipip install -r requirements.txt

